# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ORBIT Autonomous Hiring OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List all job roles
      responses:
        "200":
          description: Jobs list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a job role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a job role
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job role
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
...
  /debates:
    get:
      operationId: listDebates
      tags: [debates]
      summary: List all AI debates
      responses:
        "200":
          description: Debates list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Debate"
    post:
      operationId: createDebate
      tags: [debates]
      summary: Trigger an AI debate for a candidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DebateInput"
      responses:
        "201":
          description: Created debate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Debate"
  /debates/{id}:
    get:
      operationId: getDebate
      tags: [debates]
      summary: Get a specific debate
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Debate detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Debate"
...
components:
  schemas:
...
    Job:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        department:
          type: string
        location:
          type: string
        description:
          type: string
        requirements:
          type: array
          items:
            type: string
        candidateCount:
          type: integer
        status:
          type: string
        createdAt:
          type: string
      required: [id, title, department, location, description, requirements, candidateCount, status, createdAt]
    JobInput:
      type: object
      properties:
        title:
          type: string
        department:
          type: string
        location:
          type: string
        description:
          type: string
        requirements:
          type: array
          items:
            type: string
      required: [title, department, location, description]
    JobUpdate:
      type: object
      properties:
        title:
          type: string
        status:
          type: string
        description:
          type: string
...
    DebateMessage:
      type: object
      properties:
        agent:
          type: string
        agentType:
          type: string
        message:
          type: string
        stance:
          type: string
        evidence:
          type: array
          items:
            type: string
        timestamp:
          type: string
      required: [agent, agentType, message, stance, timestamp, evidence]
    Debate:
      type: object
      properties:
        id:
          type: integer
        candidateId:
          type: integer
        candidateName:
          type: string
        jobTitle:
          type: string
        messages:
          type: array
          items:
            $ref: "#/components/schemas/DebateMessage"
        verdict:
          type: string
        confidenceScore:
          type: number
        technicalScore:
          type: number
        cultureScore:
          type: number
        growthScore:
          type: number
        riskScore:
          type: number
        summary:
          type: string
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - candidateId
        - candidateName
        - jobTitle
        - messages
        - verdict
        - confidenceScore
        - technicalScore
        - cultureScore
        - growthScore
        - riskScore
        - summary
        - status
        - createdAt
    DebateInput:
      type: object
      properties:
        candidateId:
          type: integer
        jobId:
          type: ["integer", "null"]
      required: [candidateId]
...