# lib/api-zod/src/generated/types/debate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ORBIT Autonomous Hiring OS API
 * OpenAPI spec version: 0.1.0
 */
import type { DebateMessage } from './debateMessage';

export interface Debate {
  id: number;
  candidateId: number;
  candidateName: string;
  jobTitle: string;
  messages: DebateMessage[];
  verdict: string;
  confidenceScore: number;
  technicalScore: number;
  cultureScore: number;
  growthScore: number;
  riskScore: number;
  summary: string;
  status: string;
  createdAt: string;
}