# lib/api-zod/src/generated/types/jobUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ORBIT Autonomous Hiring OS API
 * OpenAPI spec version: 0.1.0
 */

export interface JobUpdate {
  title?: string;
  status?: string;
  description?: string;
}